refactor: move About into a Settings tab
Removes the standalone About window and its menu wiring; About is now
a tab in the Settings window. Drops the tagline and adds a
"© 2026 Ali Elsokary. All rights reserved." line pinned at the bottom.
